compose dry run
